Move Flight Fixture to use Middleware instead of WebDevServer
This lets us put it in the same server that would be serving this content
in a more real world scenario.

I also de-CRA:ified this a bit by simplifying pieces we don't need.
