+++ fixtures/flight/config/modules.js
@@ -3,7 +3,7 @@
 const fs = require('fs');
 const path = require('path');
 const paths = require('./paths');
-const chalk = require('react-dev-utils/chalk');
+const chalk = require('chalk');
 const resolve = require('resolve');
 
 /**
