+++ fixtures/flight/server/package.json
@@ -1,4 +1,3 @@
 {
-  "type": "commonjs",
-  "main": "./cli.js"
+  "type": "commonjs"
 }
